+++ app/src/js/components/Granules/overview.js
@@ -34,10 +34,11 @@ import statusOptions from '../../utils/status';
 import _config from '../../config';
 import { strings } from '../locale';
 import { workflowOptionNames } from '../../selectors';
-import { window, document } from '../../utils/browser';
+import { window } from '../../utils/browser';
 import ListFilters from '../ListActions/ListFilters';
 import Breadcrumbs from '../Breadcrumbs/Breadcrumbs';
 import pageSizeOptions from '../../utils/page-size';
+import { downloadFile } from '../../utils/download-file';
 
 const { updateInterval } = _config;
 
@@ -135,14 +136,8 @@ class GranulesOverview extends React.Component {
       const { granuleCSV } = this.props;
       const { data } = granuleCSV;
       const csvData = new Blob([data], { type: 'text/csv' });
-
-      const link = document.createElement('a');
-      link.setAttribute('download', 'granules.csv');
       const url = window.URL.createObjectURL(csvData);
-      link.href = url;
-      document.body.appendChild(link);
-      link.click();
-      link.parentNode.removeChild(link);
+      downloadFile(url, 'granules.csv');
     });
   }
 
@@ -170,7 +165,7 @@ class GranulesOverview extends React.Component {
         </section>
         <section className='page__section'>
           <div className='heading__wrapper--border'>
-            <h2 className='heading--medium heading--shared-content with-description'>{strings.granules} <span className='num--title'>{count ? ` ${tally(count)}` : 0}</span></h2>
+            <h2 className='heading--medium heading--shared-content with-description'>{strings.granules} <span className='num-title'>{count ? ` ${tally(count)}` : 0}</span></h2>
             <a className='csv__download button button--small button--download button--green form-group__element--right'
               id='download_link'
               onClick={this.downloadGranuleCSV}
@@ -237,7 +232,6 @@ GranulesOverview.propTypes = {
   stats: PropTypes.object,
   dispatch: PropTypes.func,
   workflowOptions: PropTypes.array,
-  location: PropTypes.object,
   config: PropTypes.object,
   granuleCSV: PropTypes.object
 };