+++ app/src/js/components/Pdr/overview.js
@@ -20,8 +20,20 @@ import _config from '../../config';
 import Dropdown from '../DropDown/dropdown';
 import pageSizeOptions from '../../utils/page-size';
 import ListFilters from '../ListActions/ListFilters';
+import Breadcrumbs from '../Breadcrumbs/Breadcrumbs';
 
 const { updateInterval } = _config;
+
+const breadcrumbConfig = [
+  {
+    label: 'Dashboard Home',
+    href: '/'
+  },
+  {
+    label: 'PDRs',
+    active: true
+  }
+];
 
 class PdrOverview extends React.Component {
   constructor () {
@@ -70,6 +82,9 @@ class PdrOverview extends React.Component {
     const overview = this.renderOverview(pdrCount);
     return (
       <div className='page__component'>
+        <section className='page__section page__section__controls'>
+          <Breadcrumbs config={breadcrumbConfig} />
+        </section>
         <section className='page__section page__section__header-wrapper'>
           <h1 className='heading--large heading--shared-content with-description'>PDR Overview</h1>
           {lastUpdated(queriedAt)}
@@ -77,7 +92,7 @@ class PdrOverview extends React.Component {
         </section>
         <section className='page__section'>
           <div className='heading__wrapper--border'>
-            <h2 className='heading--medium heading--shared-content with-description'>All PDRs <span className='num--title'>{count ? ` ${tally(count)}` : 0}</span></h2>
+            <h2 className='heading--medium heading--shared-content with-description'>All PDRs <span className='num-title'>{count ? ` ${tally(count)}` : 0}</span></h2>
           </div>
           <List
             list={list}