+++ app/src/js/components/Workflows/overview.js
@@ -27,7 +27,7 @@ const WorkflowOverview = ({ workflows }) => {
       </section>
       <section className='page__section'>
         <div className='heading__wrapper--border'>
-          <h2 className='heading--medium heading--shared-content with-description'>All Workflows <span className='num--title'>{count ? ` ${tally(count)}` : 0}</span></h2>
+          <h2 className='heading--medium heading--shared-content with-description'>All Workflows <span className='num-title'>{count ? ` ${tally(count)}` : 0}</span></h2>
         </div>
         {/* Someone needs to define the search parameters for workflows, e.g. steps, collections, granules, etc. } */}
         <div className='filters'>
@@ -46,7 +46,6 @@ const WorkflowOverview = ({ workflows }) => {
           action={listWorkflows}
           tableColumns={tableColumns}
           query={{}}
-          sortId='name'
           rowId='name'
         />
       </section>